Currencies, Update Fix, Settings Added
Reshaped Form2 to deal with currencies, removed api fields as they are
not in use yet.

Changed Application ID in setup, allowing updates (this tricks
installshield into thinking this is a Major Update...there has to be a
better way)

Added an update path.

Removed hard limits on digit length for the LCD - this allows for yen
and future bitcoin growth.

Removed currency symbols - displays what Mt Gox provides.

Decoupled target url for the api - in theory it'll be slightly easier to
add other exchanges in the future.

Changed menu from 'info' to 'about.'

Added application settings - currency preference is stored there.

NOTE: there was some wonkiness with the .sln file between the last
commit and this one. Something may have broke.